// lib/types/product.ts
/**
 * Product-related types
 */

export interface ProductRating {
  rate: number;
  count: number;
}

export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: ProductRating;
}

export type ProductCardProps = {
  product: Product;
};